rename and move classes from util directory
